Implement file revving for cache busting

+++ generate-html.sh
@@ -1,5 +1,15 @@
 #!/usr/bin/env sh
 
+###################
+#
+# Tools required:
+# - pandoc
+# - sha256sum
+# - sed
+# - cut
+#
+###################
+
 if ! command -v pandoc > /dev/null 2>&1; then
   echo "Make sure pandoc is installed, then try again"
   exit 1
@@ -11,10 +21,29 @@ slugify() (
   printf '%s%s' "${1%"$filename"}" "$slugified_name"
 )
 
+# Generate a hash of styles.css to allow for long-term caching and simple invalidation.
+if command -v sha256sum >/dev/null 2>&1; then
+  styles_hash="$(sha256sum generator-files/styles.css | cut -c1-16)"
+elif command -v shasum; then
+  styles_hash="$(shasum -a 256 generator-files/styles.css | cut -c1-16)"
+elif command -v openssl; then
+  styles_hash="$(openssl sha256 generator-files/styles.css | cut -d ' ' -f 2 | cut -c1-16)"
+else
+  printf "Could not generate sha256 hash of styles.css, make sure sha256sum, shasum, or openssl are available on your system\n"
+  exit 1
+fi
+
+if [ -z "$styles_hash" ]; then
+  printf "sha256 hash was empty, an unknown error occurred"
+  exit 1
+fi
+
+# Cleanup files from previous builds
 rm -r dist
 mkdir dist
 
-cp generator-files/{styles.css,favicon.ico} dist/
+cp generator-files/styles.css "dist/styles-${styles_hash}.css"
+cp generator-files/favicon.ico dist/
 cp generator-files/index.start.html dist/index.html
 
 for markdown_file in recipes/*.md; do
@@ -33,6 +62,7 @@ for markdown_file in recipes/*.md; do
       -o "$output_name/index.html" \
       --data-dir ./
 
+    # Add a link for this recipe to our index page
     printf '<li class="Recipes-item"><a href="./%s/">%s</a></li>\n' \
       "$(basename "$output_name")" \
       "$(basename "${markdown_file%.*}")" \
@@ -40,10 +70,14 @@ for markdown_file in recipes/*.md; do
   fi
 done
 
+# Finish constructing the index page
 cat generator-files/index.end.html >> dist/index.html
 
+# Add the first 16 characters from the styles.css hash to its filename
+sed -i "s/STYLESDOTCSS_HASH/${styles_hash}/g" dist/*.html dist/**/*.html
+
 # Try to compress the files ahead of time so the webserver can do less work
-for dist_file in dist/{*.*,**/*.*}; do
+for dist_file in dist/*.* dist/**/*.*; do
   if [ -e "$dist_file" ] && [ ! -d "$dist_file" ]; then
     if command -v brotli > /dev/null 2>&1; then
       brotli --keep --best "$dist_file"
